[JENKINS-73805] implement split
