+++ core/src/main/java/hudson/AboutJenkins.java
@@ -4,6 +4,7 @@ import edu.umd.cs.findbugs.annotations.NonNull;
 import hudson.model.ManagementLink;
 import hudson.security.Permission;
 import java.net.URL;
+import java.util.List;
 import jenkins.model.Jenkins;
 import org.jenkinsci.Symbol;
 import org.kohsuke.accmod.Restricted;
@@ -51,5 +52,13 @@ public class AboutJenkins extends ManagementLink {
     @Override
     public Category getCategory() {
         return Category.STATUS;
+    }
+
+    public static AboutPageDecorator getAboutPageDecorator() {
+        return AboutPageDecorator.first();
+    }
+
+    public static List<AboutPageDecorator> getAboutPageDecorators() {
+        return AboutPageDecorator.all();
     }
 }
