+++ core/src/main/resources/hudson/AboutJenkins/about-static.jelly
@@ -1,0 +1,52 @@
+<!--
+The MIT License
+
+Copyright (c) 2023, CloudBees, Inc.
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in
+all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+THE SOFTWARE.
+-->
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:l="/lib/layout">
+  <tr>
+    <td>
+      <a class="jenkins-table__link" href="https://github.com/jenkins-contrib-themes/jenkins-core-theme">
+        Jenkins Contrib Themes
+      </a>
+    </td>
+    <td>
+      <a class="jenkins-table__link" href="https://github.com/afonsof">Afonso Franca</a>
+    </td>
+    <td>
+      <a class="jenkins-table__link" href="https://opensource.org/licenses/MIT">MIT License</a>
+    </td>
+  </tr>
+  <tr>
+    <td>
+      <a class="jenkins-table__link" href="https://ionic.io/ionicons">Ionicons</a>
+    </td>
+    <td>
+      <a class="jenkins-table__link" href="https://github.com/ionic-team">Ionic</a>
+    </td>
+    <td>
+      <a class="jenkins-table__link" href="https://github.com/ionic-team/ionicons/blob/master/LICENSE">MIT
+        License
+      </a>
+    </td>
+  </tr>
+</j:jelly>
