+++ core/src/main/resources/hudson/AboutJenkins/index.properties
@@ -20,6 +20,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
+jenkins.name=Jenkins
 about=About Jenkins {0}
 blurb=The leading open source automation server which enables developers around the world to reliably build, test, and deploy their software.
